feat: implement vector-based matching and hybrid skill extraction
- Database: Added embedding columns to Resume and Job models to persist vector embeddings and eliminate redundant runtime compute.
- Matching Engine: Replaced job_matcher with vector_matcher and updated match_routes to calculate cosine similarity on pre-computed embeddings.
- NLP/Extraction: Upgraded embedding_skill_extractor to use a hybrid approach (exact keyword matching + semantic scoring) to prevent signal dilution.
- ATS Engine: Refactored calculate_ats_score to focus on structural and formatting heuristics.
- Cleanup: Removed deprecated matching_routes.py and standardized embedding generation return types.

+++ backend/app/services/ats/ats_calculator.py
@@ -62,12 +62,12 @@ def calculate_ats_score(resume_text: str, job_text: str, missing_skills=None):
     )
 
     return {
-        "ats_score": round(final_score, 2),
+        "ats_score": float(round(final_score, 2)),
         "breakdown": {
-            "keyword_match": round(kw, 2),
-            "semantic_similarity": round(sem, 2),
-            "skill_match": round(skill_score, 2),
-            "structure_score": round(struct, 2),
+            "keyword_match": float(round(kw, 2)),
+            "semantic_similarity": float(round(sem, 2)),
+            "skill_match": float(round(skill_score, 2)),
+            "structure_score": float(round(struct, 2)),
         },
         "missing_skills": missing_skills or []
     }
